removed a bunch of weird code
