+++ datatypes.h
@@ -1156,9 +1156,4 @@ public:
 
 Q_DECLARE_METATYPE(ENCODER_DETECT_RES)
 
-typedef enum {
-    TCP_HUB_CONNECT_TO_VESC = 0,
-    TCP_HUB_LIST_VESCS,
-} TCP_HUB_CONNECTION_CMD;
-
 #endif // DATATYPES_H
