--- tcpconnectedvesc.cpp
@@ -1,16 +1,0 @@
-#include "tcpconnectedvesc.h"
-
-TcpConnectedVesc::TcpConnectedVesc(QObject *parent) : QObject{parent}
-{
-
-}
-
-void TcpConnectedVesc::setTcpSocket(QTcpSocket *tcpSocket)
-{
-    mTcpSocket = tcpSocket;
-}
-
-QTcpSocket *TcpConnectedVesc::tcpSocket() const
-{
-    return mTcpSocket;
-}
