+++ tcphub.cpp
@@ -6,6 +6,10 @@ TcpHub::TcpHub(QObject *parent)
 {
     mTcpHubServer = new QTcpServer(this);
     connect(mTcpHubServer, SIGNAL(newConnection()), this, SLOT(newTcpHubConnection()));
+
+    mPeriodic = new QTimer(this);
+    connect(mPeriodic, SIGNAL(timeout()), this, SLOT(periodic()));
+    mPeriodic->start(1000);
 }
 
 void TcpHub::addVesc(QString id, TcpConnectedVesc *vesc)
@@ -29,28 +33,11 @@ void TcpHub::newTcpHubConnection()
     QHostAddress addr = socket->peerAddress();
 
     qDebug() << "new connection to hub: " << addr.toString();
-
-    socket->write("VESC_TOOL\n");
-    socket->write(mVescIF->getTcpHubVescID().toLocal8Bit());
-    socket->write(":");
-    socket->write(mVescIF->getTcpHubVescPass().toLocal8Bit());
-    socket->write("\n");
 }
 
-void TcpHub::newVescConnection(QTcpSocket *socket)
+void TcpHub::periodic()
 {
-    TcpConnectedVesc *vesc = nullptr;
-    QString addr = socket->peerAddress().toString();
-    if (mTcpConnectedVescs.contains(addr)) {
-        vesc = mTcpConnectedVescs[addr];
-        socket->setParent(vesc);
-        vesc->setTcpSocket(socket);
-    } else {
-        vesc = new TcpConnectedVesc(this);
-        socket->setParent(vesc);
-        vesc->setTcpSocket(socket);
-        mTcpConnectedVescs.insert(addr,vesc);
-    }
+    qDebug("periodic");
 }
 
 VescInterface *TcpHub::vescIF() const
